+++ src/main/java/jamierhys/comtech/common/items/tools/ToolObsidianCutter.java
@@ -1,0 +1,17 @@
+package jamierhys.comtech.common.items.tools;
+
+import jamierhys.comtech.client.tabs.CreativeTab;
+import jamierhys.comtech.common.items.tools.materials.ModMaterials;
+import jamierhys.comtech.main.ComTech;
+import net.minecraft.item.ItemPickaxe;
+
+public class ToolObsidianCutter extends ItemPickaxe {
+    public ToolObsidianCutter() {
+        super(ModMaterials.ToolMaterialObsidian);
+
+        this.setCreativeTab(CreativeTab.CT_TAB);
+        this.setMaxStackSize(1);
+        this.setUnlocalizedName(ComTech.Reference.Resource.Prefix + "tool_obsidiancutter");
+        this.setRegistryName(ComTech.Reference.ModInfo.ModID, "tool_obsidiancutter");
+    }
+}