+++ CHANGELOG.md
@@ -5,10 +5,15 @@ This mod adheres to [Schemantic Versioning](https://simver.org)
 
 ## [unreleased]
 
+## [0.0.1]
+
 | Added                                                        
 |--------------------------------------------------------------
 | Added dust items
 | Added ingot items
+| Added rock items
+| Added tools items
+| Added Usable items
 
 | Changed                                                      
 |--------------------------------------------------------------
@@ -20,3 +25,4 @@ This mod adheres to [Schemantic Versioning](https://simver.org)
 |--------------------------------------------------------------
 
 [unreleased]:https://github.com/JamieRhys/Comtech
+[0.0.1]: